Support finding datalists in specific sites.
You can configure the control to determine in which site it should look
for datalists. If unconfigured, it searches across the entire
repository, as before.
Use <control-param name="parentSite">foo</control-param> to search in
site foo.
Use <control-param name="parentSiteProp">cm:parentSite</control-param>
to use cm:parentSite as the property that contains the site name "foo"
to configure it per-folder.